+++ app/src/main/java/com/ala158/magicpantry/data/Ingredient.kt
@@ -25,7 +25,7 @@ data class Ingredient(
     var isNotify: Boolean = false,
 
     @ColumnInfo(name = "notify_threshold")
-    var notifyThreshold: Int = 0,
+    var notifyThreshold: Double = 0.0,
 ) {
     constructor(name: String, amount: Double, unit: String, price: Double) : this() {
         this.name = name
@@ -34,7 +34,7 @@ data class Ingredient(
         this.price = price
     }
 
-    constructor(name: String, amount: Double, unit: String, price: Double, notifyThreshold: Int): this() {
+    constructor(name: String, amount: Double, unit: String, price: Double, notifyThreshold: Double): this() {
         this.name = name
         this.amount = amount
         this.unit = unit
